+++ packages/react/src/select/backdrop/SelectBackdrop.tsx
@@ -7,12 +7,12 @@ import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import { HTMLElementType } from '../../utils/proptypes';
 import { useSelectRootContext } from '../root/SelectRootContext';
 import { useSelectBackdrop } from './useSelectBackdrop';
-import { popupOpenStateMapping } from '../../utils/popupOpenStateMapping';
+import { popupStateMapping } from '../../utils/popupStateMapping';
 import type { CustomStyleHookMapping } from '../../utils/getStyleHookProps';
 import type { TransitionStatus } from '../../utils/useTransitionStatus';
 
 const customStyleHookMapping: CustomStyleHookMapping<SelectBackdrop.State> = {
-  ...popupOpenStateMapping,
+  ...popupStateMapping,
   transitionStatus(value): Record<string, string> | null {
     if (value === 'entering') {
       return { 'data-starting-style': '' };
