+++ packages/react/src/select/popup/SelectPopup.tsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { FloatingFocusManager, type Side } from '@floating-ui/react';
 import type { BaseUIComponentProps } from '../../utils/types';
 import { useSelectRootContext } from '../root/SelectRootContext';
-import { popupOpenStateMapping } from '../../utils/popupOpenStateMapping';
+import { popupStateMapping } from '../../utils/popupStateMapping';
 import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import { useForkRef } from '../../utils/useForkRef';
 import type { CustomStyleHookMapping } from '../../utils/getStyleHookProps';
@@ -13,7 +13,7 @@ import type { TransitionStatus } from '../../utils/useTransitionStatus';
 import { useSelectPositionerContext } from '../positioner/SelectPositionerContext';
 
 const customStyleHookMapping: CustomStyleHookMapping<SelectPopup.State> = {
-  ...popupOpenStateMapping,
+  ...popupStateMapping,
   transitionStatus(value): Record<string, string> | null {
     if (value === 'entering') {
       return { 'data-starting-style': '' };
